Parallelize e2e exception tests
