--- source/Tests.Library/E2EScenarios/FullE2EExceptionCases.cs
@@ -1,252 +1,0 @@
-﻿using Sailfish;
-using Shouldly;
-using System.Linq;
-using System.Threading.Tasks;
-using Tests.Common.Utils;
-using Tests.E2E.ExceptionHandling;
-using Tests.E2E.ExceptionHandling.Tests;
-using Xunit;
-
-namespace Tests.Library.E2EScenarios;
-
-public class FullE2EExceptionCases
-{
-    [Fact]
-    public async Task GlobalSetupExceptionsAreHandled()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(GlobalSetupExceptionIsHandled))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("Global Setup Exception");
-    }
-
-    [Fact]
-    public async Task MethodSetupExceptionsAreHandled()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(MethodSetupExceptionIsHandled))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().InnerException?.Message.ShouldBe("Method Setup Exception");
-    }
-
-    [Fact]
-    public async Task IterationSetupExceptionsAreHandled()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(IterationSetupExceptionIsHandled))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("Iteration Setup Exception");
-    }
-
-    [Fact]
-    public async Task IterationTeardownExceptionsAreHandled()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(IterationTeardownExceptionIsHandled))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("Iteration Teardown Exception");
-    }
-
-    [Fact]
-    public async Task MethodTeardownExceptionsAreHandled()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(MethodTeardownExceptionIsHandled))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("Method Teardown Exception");
-    }
-
-    [Fact]
-    public async Task GlobalTeardownExceptionsAreHandled()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(GlobalTeardownExceptionIsHandled))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("Global Teardown Exception");
-    }
-
-    [Fact]
-    public async Task MultipleLifecycleExceptionsAreHandledWithMethodTeardownSurfacing()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(MethodTeardownExceptionComesFirst))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("Method Teardown Exception");
-    }
-
-    [Fact]
-    public async Task MultipleLifecycleExceptionsAreHandledWithIterationSetupSurfacing()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(IterationSetupExceptionComesFirst))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(2);
-        result.Exceptions.ToList()[0].Message.ShouldBe("Iteration Setup Exception");
-        result.Exceptions.ToList()[1].Message.ShouldBe("Iteration Setup Exception");
-    }
-
-    [Fact]
-    public async Task OnlySailfishMethodThrows()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(OnlyTheSailfishMethodThrows))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.IsValid.ShouldBeFalse();
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("Sailfish Method Exception");
-    }
-
-    [Fact]
-    public async Task MultipleInjectionsOnAMethodThrows()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(MultipleInjectionsOnAsyncMethod))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.IsValid.ShouldBeFalse();
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions.Count().ShouldBe(1);
-        result.Exceptions.Single().Message.ShouldBe("The 'MainMethod' method in class 'MultipleInjectionsOnAsyncMethod' may only receive a single 'CancellationToken' parameter");
-    }
-
-    [Fact]
-    public async Task WhenTestExceptionOccursHandlersAreOnlyGivenRealData()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(SailfishMethodException))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.IsValid.ShouldBeFalse();
-
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions?.Count().ShouldBe(1);
-        result.ExecutionSummaries
-            .SelectMany(x => x.CompiledTestCaseResults.Select(c => c.PerformanceRunResult))
-            .Count(x => x is null)
-            .ShouldBe(1);
-    }
-
-    [Fact]
-    public async Task WhenOneTestCaseThrowsTheRemainingTestCasesStillExecute()
-    {
-        var runSettings = RunSettingsBuilder.CreateBuilder()
-            .WithLocalOutputDirectory(Some.RandomString())
-            .WithTestNames(nameof(TestsRunIfException))
-            .ProvidersFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .TestsFromAssembliesContaining(typeof(E2ETestExceptionHandlingProvider))
-            .Build();
-        var result = await SailfishRunner.Run(runSettings);
-
-        result.IsValid.ShouldBeFalse();
-
-        result.Exceptions.ShouldNotBeNull();
-        result.Exceptions?.Count().ShouldBe(1);
-
-        result.ExecutionSummaries.First().CompiledTestCaseResults.First().TestCaseId.ShouldNotBeNull();
-        result.ExecutionSummaries
-            .SelectMany(x => x.CompiledTestCaseResults.Select(c => c.PerformanceRunResult))
-            .Count(x => x is null)
-            .ShouldBe(1);
-
-        result.ExecutionSummaries
-            .SelectMany(x => x.CompiledTestCaseResults.Select(c => c.PerformanceRunResult))
-            .Count(x => x is not null)
-            .ShouldBe(2);
-    }
-}
